Fixed Time Update
Currently updates every 60 seconds. (Time does not sync on initial start.)

+++ RGBW_controller_and_thermometer/universalFunc.h
@@ -1,3 +1,4 @@
+
 //Math Stuff
 
 bool between(int _target, int _bottom, int _top) {
@@ -39,7 +40,7 @@ void timerSetup() {
   }
 }
 
-bool timer(int _interval,int _id) { //_interval in millis, _id in countData
+bool timer(unsigned long _interval,int _id) { //_interval in millis, _id in countData
   if (millisCount(1,_id) >= _interval) {
     millisCount(0,_id);
     return true;
